Foxed bad plugin - removed Changed timeouts in site-reliability checks Fixed a bad date in teh redirect for one okay, that looks good and that's great, but let's put some comments into why we have to do that into that case_preserving_permalinks.rb file. And let's update the documentation on the project too while we're at it.
That plug-in we wrote—that plug-in because the Apple file system is case insensitive (or at least it wasn't built case sensitive), but we have file names that have mixed case and Apple supports file names in mixed case. This is what was happening: we needed to have them consistent across the development and the production.

Checkl was not smart enough to handle it by saying case sensitive; it just did it—wasn't smart enough. So we need to make some notes as to why this happened.

So when we come back next time, we go, "Ah, that's the problem." Okay? So I'm glad. This looks like it works. I'm going to try and make the commit. You start looking at the documentation that needs to be done. I'm going to just push the commit right now because that's all we've got to do  of the LogGPT pages.